Moved color config to plugin file to fix sequencing
Prior to this, nord's colours would load after my colours.lua setup
meaning I couldn't override. Updated to use packer config so it will
load after the plugin executes
