+++ dot_config/nvim/lua/tap/plugins/nord.lua
@@ -1,0 +1,36 @@
+local highlight = require("tap.utils").highlight
+local augroup = require("tap.utils").augroup
+local color = require("tap.utils").color
+
+local function apply_user_highlights()
+    highlight('Search', {
+        guibg = color({dark = "nord9_gui", light = "blue2"}),
+        guifg = color({dark = "nord0_gui", light = "bg"}),
+        gui = "NONE"
+    })
+    highlight('IncSearch', {
+        guibg = color({dark = "nord9_gui", light = "blue2"}),
+        guifg = color({dark = "nord0_gui", light = "bg"}),
+        gui = "NONE"
+    })
+
+    -- Treesitter overrides
+    highlight('TSInclude', {gui = 'italic', cterm = 'italic'})
+    highlight('TSOperator', {gui = 'italic', cterm = 'italic'})
+    highlight('TSKeyword', {gui = 'italic', cterm = 'italic'})
+
+    highlight('gitmessengerHeader', {link = 'Identifier'})
+    highlight('gitmessengerHash', {link = 'Comment'})
+    highlight('gitmessengerHistory', {link = 'Constant'})
+    highlight('gitmessengerPopupNormal', {link = 'CursorLine'})
+end
+
+augroup("ExplorerHighlights", {
+    {
+        events = {"VimEnter", "ColorScheme"},
+        targets = {"*"},
+        command = apply_user_highlights
+    }
+})
+
+apply_user_highlights()
